pyproject.toml
[project]
name = "llm_summarize"
version = "0.1.0"
description = "A telegram bot that creates text and TTS summaries of RSS feed entries using LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "black>=24.10.0",
    "feedparser>=6.0.11",
    "flake8>=7.1.1",
    "kokoro>=0.7.16",
    "ollama>=0.4.2",
    "openai>=1.57.4",
    "pip>=25.0.1",
    "psycopg[binary]>=3.2.3",
    "pydantic>=2.10.3",
    "python-telegram-bot[job-queue,rate-limiter]>=21.9",
    "requests>=2.32.3",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.37",
]
